Split services utilities.js
