+++ src/services/debug.js
@@ -1,0 +1,9 @@
+
+export default function (msg) {
+  return hook => {
+    console.log(`* ${msg || ''}\ntype:${hook.type}, method: ${hook.method}`);
+    if (hook.data) { console.log('data:', hook.data); }
+    if (hook.params && hook.params.query) { console.log('query:', hook.params.query); }
+    if (hook.result) { console.log('result:', hook.result); }
+  };
+}
